app/requirements.txt
Django~=2.1
psycopg2-binary
sorl-thumbnail
Pillow
django-bootstrap4
